# generated/node/src/routes/partners.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listPartner } from '../functions/list-partner.function.js';
import { registerPartner } from '../functions/register-partner.function.js';
import { findPartnerById } from '../functions/find-partner-by-id.function.js';
import { updatePartner } from '../functions/update-partner.function.js';
import { approvePartner } from '../functions/approve-partner.function.js';
import { deletePartner } from '../functions/delete-partner.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const partnersRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/partners', async (request, reply) => {
    try {
      const result = listPartner(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listPartner failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/partners/register', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = registerPartner(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'registerPartner failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/partners/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findPartnerById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findPartnerById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/partners/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updatePartner(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updatePartner failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/partners/:id/approve', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = approvePartner(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'approvePartner failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/partners/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = deletePartner(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'deletePartner failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};