# generated/node/src/routes/patterns.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listPatterns } from '../functions/list-patterns.function.js';
import { findPatternById } from '../functions/find-pattern-by-id.function.js';
import { detectPatterns } from '../functions/detect-patterns.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const patternsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/patterns', async (request, reply) => {
    try {
      const result = listPatterns(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listPatterns failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/patterns/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findPatternById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findPatternById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/patterns/detect', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = detectPatterns(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'detectPatterns failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};