# generated/node/src/routes/profileSessions.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listProfileSessions } from '../functions/list-profile-sessions.function.js';
import { findProfileSessionById } from '../functions/find-profile-session-by-id.function.js';
import { startProfile } from '../functions/start-profile.function.js';
import { stopProfile } from '../functions/stop-profile.function.js';
import { listPerformancePatterns } from '../functions/list-performance-patterns.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const profileSessionsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/profile-sessions', async (request, reply) => {
    try {
      const result = listProfileSessions(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listProfileSessions failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/profile-sessions/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findProfileSessionById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findProfileSessionById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/profile-sessions/start', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = startProfile(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'startProfile failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/profile-sessions/:id/stop', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = stopProfile(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'stopProfile failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/profile-sessions/:id/patterns', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = listPerformancePatterns(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listPerformancePatterns failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};