# generated/node/src/routes/solution-channels.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listSolutionChannel } from '../functions/list-solution-channel.function.js';
import { publishToChannel } from '../functions/publish-to-channel.function.js';
import { findSolutionChannelById } from '../functions/find-solution-channel-by-id.function.js';
import { updateSolutionChannel } from '../functions/update-solution-channel.function.js';
import { unpublishFromChannel } from '../functions/unpublish-from-channel.function.js';
import { testSolutionChannel } from '../functions/test-solution-channel.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const solutionChannelsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/solution-channels', async (request, reply) => {
    try {
      const result = listSolutionChannel(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listSolutionChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solution-channels', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = publishToChannel(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'publishToChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solution-channels/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findSolutionChannelById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findSolutionChannelById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.put('/api/solution-channels/:id', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = updateSolutionChannel(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'updateSolutionChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.delete('/api/solution-channels/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = unpublishFromChannel(db, id);
      return reply.status(204).send();
    } catch (err: unknown) {
      request.log.error(err, 'unpublishFromChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solution-channels/:id/test', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = testSolutionChannel(db, body);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'testSolutionChannel failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};