# generated/node/src/routes/solution-runs.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { runSolution } from '../functions/run-solution.function.js';
import { listSolutionRuns } from '../functions/list-solution-runs.function.js';
import { findSolutionRunById } from '../functions/find-solution-run-by-id.function.js';
import { getSolutionMetrics } from '../functions/get-solution-metrics.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const solutionRunsRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.post('/api/solution-runs/run', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = runSolution(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'runSolution failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solution-runs', async (request, reply) => {
    try {
      const result = listSolutionRuns(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listSolutionRuns failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solution-runs/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findSolutionRunById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findSolutionRunById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solution-runs/metrics', async (request, reply) => {
    try {
      const result = getSolutionMetrics(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'getSolutionMetrics failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};