# generated/node/src/routes/solution-templates.routes.ts
// @generated by flusk-lang — DO NOT EDIT

import { type FastifyInstance } from 'fastify';
import type Database from 'better-sqlite3';
import { listSolutionTemplate } from '../functions/list-solution-template.function.js';
import { findSolutionTemplateById } from '../functions/find-solution-template-by-id.function.js';
import { createSolutionTemplate } from '../functions/create-solution-template.function.js';

interface AppWithDb extends FastifyInstance { db: Database.Database; }

export const solutionTemplatesRoutes = async (app: FastifyInstance): Promise<void> => {
  const { db } = app as AppWithDb;

  app.get('/api/solution-templates', async (request, reply) => {
    try {
      const result = listSolutionTemplate(db);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'listSolutionTemplate failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.get('/api/solution-templates/:id', async (request, reply) => {
    try {
      const { id } = request.params as Record<string, string>;
      const result = findSolutionTemplateById(db, id);
      if (result === null || result === undefined) {
        return reply.status(404).send({ error: 'Not found' });
      }
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'findSolutionTemplateById failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

  app.post('/api/solution-templates', async (request, reply) => {
    try {
      const body = request.body as Record<string, unknown>;
      const result = createSolutionTemplate(db, body);
      return reply.send(result);
    } catch (err: unknown) {
      request.log.error(err, 'createSolutionTemplate failed');
      return reply.status(500).send({ error: 'Internal server error' });
    }
  });

};