# generated/openapi.yaml
openapi: 3.1.0
info:
  title: Flusk API
  version: 0.1.0
paths:
...
  /api/partners/register:
    post:
      operationId: registerPartner
      tags:
        - partners
      summary: Register a new partner
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Partner'
  /api/partners/:id:
    get:
      operationId: findPartnerById
      tags:
        - partners
      summary: Get partner by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
    put:
      operationId: updatePartner
      tags:
        - partners
      summary: Update partner
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Partner'
    delete:
      operationId: deletePartner
      tags:
        - partners
      summary: Suspend/delete partner
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
  /api/partners/:id/approve:
    post:
      operationId: approvePartner
      tags:
        - partners
      summary: Approve a pending partner
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Partner'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Partner'
...
  /api/patterns/detect:
    post:
      operationId: detectPatterns
      tags:
        - patterns
      summary: Run pattern detection
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pattern'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pattern'
...
  /api/profile-sessions/start:
    post:
      operationId: startProfile
      tags:
        - profileSessions
      summary: Start a profiling session
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSession'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSession'
  /api/profile-sessions/:id/stop:
    post:
      operationId: stopProfile
      tags:
        - profileSessions
      summary: Stop a profiling session
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSession'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSession'
...
  /api/solution-channels/:
    get:
      operationId: listSolutionChannel
      tags:
        - solution-channels
      summary: List channels for a solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
    post:
      operationId: publishToChannel
      tags:
        - solution-channels
      summary: Publish solution to a channel
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolutionChannel'
  /api/solution-channels/:id:
    get:
      operationId: findSolutionChannelById
      tags:
        - solution-channels
      summary: Get channel by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
    put:
      operationId: updateSolutionChannel
      tags:
        - solution-channels
      summary: Update channel config
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolutionChannel'
    delete:
      operationId: unpublishFromChannel
      tags:
        - solution-channels
      summary: Remove channel connection
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
  /api/solution-channels/:id/test:
    post:
      operationId: testSolutionChannel
      tags:
        - solution-channels
      summary: Send a test message through this channel
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionChannel'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolutionChannel'
  /api/solution-runs/run:
    post:
      operationId: runSolution
      tags:
        - solution-runs
      summary: Execute a solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionRun'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolutionRun'
...
  /api/solution-templates/:
    get:
      operationId: listSolutionTemplate
      tags:
        - solution-templates
      summary: List public templates
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionTemplate'
    post:
      operationId: createSolutionTemplate
      tags:
        - solution-templates
      summary: Create a template from a solution
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolutionTemplate'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolutionTemplate'
...
components:
  schemas:
...
    Partner:
      type: object
      properties:
        name:
          type: string
          description: Partner name
        email:
          type: string
          description: Contact email
        website:
          type: string
          description: Partner website
        status:
          $ref: '#/components/schemas/enum'
          description: Partner status
          enum:
            - pending
            - approved
            - suspended
        apiKey:
          type: string
          description: API key for partner access
        tier:
          $ref: '#/components/schemas/enum'
          description: Partner tier
          enum:
            - free
            - standard
            - premium
        organizationId:
          type: string
          description: Associated organization
      required:
        - name
        - email
        - website
        - status
        - apiKey
        - tier
        - organizationId
      description: External partner/integrator in the Flusk ecosystem
    Pattern:
      type: object
      properties:
        promptHash:
          type: string
          description: Hash of the repeated prompt
        model:
          type: string
          description: Model used
        occurrences:
          type: integer
          description: Number of times this pattern appeared
        totalCostUsd:
          type: number
          description: Total cost of all occurrences
        avgTokens:
          type: number
          description: Average tokens per occurrence
        firstSeenAt:
          type: string
          description: When first detected
        lastSeenAt:
          type: string
          description: When last seen
        samplePrompt:
          type: string
          description: Example prompt text (truncated)
        recommendation:
          $ref: '#/components/schemas/enum'
          description: Suggested action
          enum:
            - cache
            - batch
            - ignore
        agentName:
          type: string
          description: Agent generating this pattern
      required:
        - promptHash
        - model
        - occurrences
        - totalCostUsd
        - avgTokens
        - firstSeenAt
        - lastSeenAt
        - samplePrompt
        - recommendation
        - agentName
      description: Detected repetitive prompt pattern
...
    ProfileSession:
      type: object
      properties:
        name:
          type: string
          description: Session name
        profileType:
          $ref: '#/components/schemas/enum'
          description: Type of profiling
          enum:
            - cpu
            - heap
            - both
        status:
          $ref: '#/components/schemas/enum'
          description: Session status
          enum:
            - running
            - completed
            - failed
        pid:
          type: integer
          description: Process ID being profiled
        durationMs:
          type: integer
          description: Profiling duration
        flameGraphPath:
          type: string
          description: Path to generated flamegraph SVG
        cpuProfilePath:
          type: string
          description: Path to raw CPU profile JSON
        heapSnapshotPath:
          type: string
          description: Path to heap snapshot
        hotspotCount:
          type: integer
          description: Number of hotspots detected
        peakHeapMb:
          type: number
          description: Peak heap usage in MB
        summary:
          type: object
          description: Summary statistics
      required:
        - name
        - profileType
        - status
        - pid
        - durationMs
        - flameGraphPath
        - cpuProfilePath
        - heapSnapshotPath
        - hotspotCount
        - peakHeapMb
        - summary
      description: Profiling session data from @platformatic/flame
...
    SolutionChannel:
      type: object
      properties:
        solutionId:
          type: string
          description: Solution this channel belongs to
        channel:
          $ref: '#/components/schemas/enum'
          description: Channel type
          enum:
            - slack
            - whatsapp
            - telegram
            - discord
            - api
            - web
        config:
          type: object
          description: Channel configuration (webhook URL, bot token, etc.)
        status:
          $ref: '#/components/schemas/enum'
          description: Channel status
          enum:
            - active
            - inactive
        messageCount:
          type: integer
          description: Total messages sent through this channel
        lastMessageAt:
          type: string
          description: Last message timestamp
      required:
        - solutionId
        - channel
        - config
        - status
        - messageCount
        - lastMessageAt
      description: Channel connection for publishing a solution (Slack, WhatsApp, etc.)
    SolutionRun:
      type: object
      properties:
        solutionId:
          type: string
          description: Solution that was executed
        organizationId:
          type: string
          description: Organization that owns this run
        input:
          type: object
          description: Input data for the run
        output:
          type: object
          description: Output data from the run
        status:
          $ref: '#/components/schemas/enum'
          description: Execution status
          enum:
            - running
            - completed
            - failed
            - timeout
        cost:
          type: number
          description: Total cost in USD
        durationMs:
          type: integer
          description: Execution duration in milliseconds
        modelUsed:
          type: string
          description: Model used for this run
        tokensIn:
          type: integer
          description: Input tokens consumed
        tokensOut:
          type: integer
          description: Output tokens generated
        error:
          type: string
          description: Error message if failed
        traceId:
          type: string
          description: Links to distributed trace
      required:
        - solutionId
        - organizationId
        - input
        - output
        - status
        - cost
        - durationMs
        - modelUsed
        - tokensIn
        - tokensOut
        - error
        - traceId
      description: Single execution of a solution with cost and performance tracking
    SolutionTemplate:
      type: object
      properties:
        name:
          type: string
          description: Template name
        description:
          type: string
          description: Template description
        category:
          $ref: '#/components/schemas/enum'
          description: Template category
          enum:
            - customer-support
            - data-analysis
            - content-generation
            - code-review
            - custom
        config:
          type: object
          description: Default solution configuration
        isPublic:
          type: boolean
          description: Whether template is publicly visible
        author:
          type: string
          description: Template author
        downloads:
          type: integer
          description: Number of times cloned
        rating:
          type: number
          description: Average user rating
      required:
        - name
        - description
        - category
        - config
        - isPublic
        - author
        - downloads
        - rating
      description: Pre-built solution template for quick start
...